ml/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --format requirements.txt --output-file requirements.txt
anyio==4.12.1
    # via httpx
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   nltk
    #   typer
    #   typer-slim
    #   underthesea
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.20.3
    # via huggingface-hub
fonttools==4.61.1
    # via matplotlib
fsspec==2026.1.0
    # via huggingface-hub
fugashi-plus==1.5.2.post1
    # via ml
gensim==4.4.0
    # via tltk
h11==0.16.0
    # via httpcore
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.3.4
    # via underthesea
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jieba==0.42.1
    # via ml
joblib==1.5.3
    # via
    #   nltk
    #   scikit-learn
    #   underthesea
jpype1==1.6.0
    # via konlpy
kiwisolver==1.4.9
    # via matplotlib
konlpy==0.6.0
    # via ml
lxml==6.0.2
    # via konlpy
markdown-it-py==4.0.0
    # via rich
matplotlib==3.10.8
    # via
    #   ml
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
nltk==3.9.2
    # via
    #   ml
    #   tltk
numpy==2.4.0
    # via
    #   contourpy
    #   gensim
    #   konlpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
packaging==26.0
    # via
    #   huggingface-hub
    #   jpype1
    #   matplotlib
pandas==2.3.3
    # via
    #   ml
    #   seaborn
pillow==12.1.0
    # via matplotlib
pygments==2.19.2
    # via rich
pymarc==5.3.1
    # via ml
pyparsing==3.3.2
    # via matplotlib
python-crfsuite==0.9.12
    # via
    #   sklearn-crfsuite
    #   underthesea
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   underthesea
regex==2026.1.15
    # via nltk
requests==2.32.5
    # via
    #   ml
    #   underthesea
rich==14.3.1
    # via typer
scikit-learn==1.8.0
    # via
    #   ml
    #   sklearn-crfsuite
    #   tltk
    #   underthesea
scipy==1.16.3
    # via
    #   gensim
    #   ml
    #   scikit-learn
seaborn==0.13.2
    # via ml
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
sklearn-crfsuite==0.5.0
    # via tltk
smart-open==7.5.0
    # via gensim
stop-words==2025.11.4
    # via ml
tabulate==0.9.0
    # via sklearn-crfsuite
threadpoolctl==3.6.0
    # via scikit-learn
tltk==1.10
    # via ml
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   ml
    #   nltk
    #   sklearn-crfsuite
    #   underthesea
typer==0.21.1
    # via ml
typer-slim==0.21.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   anyio
    #   huggingface-hub
    #   typer
    #   typer-slim
tzdata==2025.3
    # via pandas
underthesea==9.1.4
    # via ml
underthesea-core==1.0.7
    # via underthesea
urllib3==2.6.3
    # via requests
wrapt==2.0.1
    # via smart-open
